Move TaskError in its own file to avoid circular deps
