+++ kitipy/__init__.py
@@ -1,5 +1,6 @@
 from .dispatcher import Dispatcher
 from .context import Context, pass_context, get_current_context, get_current_executor
+from .exceptions import TaskError
 from .executor import Executor, InteractiveWarningPolicy
 from .groups import Task, Group, RootCommand, StackGroup, StageGroup, root, task, group
 from .utils import append_cmd_flags, load_config_file, normalize_config, set_up_file_transfer_listeners, wait_for
@@ -16,6 +17,9 @@ __all__ = [
     'pass_context',
     'get_current_context',
     'get_current_executor',
+
+    # from exceptions module
+    'TaskError',
 
     # from executor module
     'Executor',
@@ -39,7 +43,6 @@ __all__ = [
     # action modules
     'ansible_actions',
     'git_actions',
-    'sphinx_actions',
 
     # other modules
     'filters',
